requirements.txt
fastapi==0.115.12
uvicorn==0.34.3
python-multipart==0.0.20
tensorflow-cpu==2.18.1
Pillow==11.2.1
numpy==1.26.4
